Fixed issues with dropDownStyle
I was having issues styles applying to the dropdown via the dropDownStyles prop. I'm not sure if it's the best solution, but I moved the styling for the dropdown from the Menu component to the TextInput component. Styles also worked if it was moved to the parent View component. However, doing so did not work if it referenced the inputLayout state. So I removed references to the layout, which is where I'm not sure if this is the best solution.

Also, I changed the ts file so that setValue had the type of Dispatch<SetStateAction<string | number | undefined>>.